// Frontend/src/interfaces/products/ProductoItem.ts
export interface ProductoItem {
    producto_id: string;
    tenant_id: string;
    name: string;
    description: string;
    price: number;
    category_id: string;
    age: number;
    gender: string;
    type: string;
    availability: string;
    imageUrl: string;
    user_id: string;
    createdAt: string; // ISO date
}