// Frontend/src/interfaces/products/ObtenerProductoResponse.ts
export interface ObtenerProductoResponse {
    category_id: string;
    createdAt: string;
    producto_id: string;
    name: string;
    gender: string;
    availability: string;
    user_id: string;
    imageUrl: string;
    tenant_id: string;
    description: string;
    price: number;
    type: string;
    age: number;
}